+++ app/components/list-dialog/list-dialog.js
@@ -133,7 +133,7 @@ class ListDialog extends AppElement {
 
         button {
           min-block-size: var(--touch-target);
-          padding-inline: var(--space-4);
+          padding-inline: var(--space-2);
           border-radius: var(--radius-sm);
           border: none;
           cursor: pointer;